camcontrol_updateTargetReticle + ecsh_creator_update to 100%
camcontrol_updateTargetReticle (dll_A6) 99.02->100: int targetReticleOverride
holds a pointer; the two `!= 0` checks emit cmpwi where target uses cmplwi.
Cast (u32) at both sites (#3).

ecsh_creator_update (shrine1C2) 99.11->100: three fixes -- (u32) cast on the
GameBit_Get result and the Obj_SetupObject pointer ret (#3 cmplwi), and
*sub < 1 -> *sub <= 0 (#69 cmpwi 0 vs 1). Byte-exact.

+++ src/main/dll/dll_A6.c
@@ -41,7 +41,7 @@ void camcontrol_updateTargetReticle(u8 *fallbackTarget, int unused2,
 
   reticle = gCamcontrolTargetReticle;
   target = fallbackTarget;
-  if (CAMCONTROL_CAMERA->targetReticleOverride != 0) {
+  if ((u32)CAMCONTROL_CAMERA->targetReticleOverride != 0) {
     target = (u8 *)CAMCONTROL_CAMERA->targetReticleOverride;
     savedReticleState = gCamcontrolTargetState;
     gCamcontrolTargetState = 3;
@@ -106,7 +106,7 @@ void camcontrol_updateTargetReticle(u8 *fallbackTarget, int unused2,
   flagsObj = *(u16 **)((u8 *)*(u32 *)(reticle + 0x7C) + (s8)reticle[0xAD] * 4);
   *(u16 *)((u8 *)flagsObj + 0x18) = (u16)(*(u16 *)((u8 *)flagsObj + 0x18) & ~8);
 
-  if (CAMCONTROL_CAMERA->targetReticleOverride != 0) {
+  if ((u32)CAMCONTROL_CAMERA->targetReticleOverride != 0) {
     gCamcontrolTargetState = (s8)savedReticleState;
     ((GameObject *)reticle)->anim.alpha = savedReticleAlpha;
   }
